% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate.R
\name{roads_four_ways}
\alias{roads_four_ways}
\title{Creation of the main axes that forms the four way direction}
\usage{
roads_four_ways(x, y)
}
\arguments{
\item{x, y}{the dimension or size of the roads from east to west (x) and from north to south(y).}
}
\value{
No return value
}
\description{
This Procedure create the Roads view from the Four way direction
this procedure is automatically call once the simulation of a particular vehicle is lunched.
}
